fix: more sample and favicon for tractors
