+++ examples/react-micro-frontends/apps/team-red/webpack.config.ts
@@ -1,7 +1,8 @@
+import { withModuleFederation } from '@nx/module-federation/webpack';
 import { withReact } from '@nx/react';
-import { withModuleFederation } from '@nx/react/module-federation';
 import { composePlugins, withNx } from '@nx/webpack';
 import { withZephyr } from 'zephyr-webpack-plugin';
+import { isModuleFederationPlugin } from 'zephyr-xpack-internal';
 
 const mfConfig = {
   name: 'team-red',
@@ -33,8 +34,33 @@ const mfConfig = {
 
 // Nx plugins for webpack.
 export default composePlugins(
-  withNx(),
+  withNx({ runtimeChunk: false }),
   withReact(),
   withModuleFederation(mfConfig, { dts: false }),
+  // Workaround for aliases while this issue
+  // -> https://github.com/nrwl/nx/issues/31346
+  // is not resolved.
+  (config) => {
+    const plugin = config.plugins?.find(isModuleFederationPlugin);
+
+    if (!plugin?._options) return config;
+    if ('config' in plugin._options) return config;
+
+    plugin._options.remotes = Object.fromEntries(
+      Object.entries(plugin._options.remotes || {}).map(([name, remote]) => [
+        mfConfig.remotes.find((nameAlias) => nameAlias === name.replace(/_/g, '-')),
+        remote,
+      ])
+    );
+
+    return config;
+  },
+  // Workaround for incomplete resolution of https://github.com/nrwl/nx/issues/31114
+  (config) => {
+    if (config.optimization) {
+      config.optimization.runtimeChunk = false;
+    }
+    return config;
+  },
   withZephyr()
 );
