+++ examples/rspack-nx-mf/apps/host/module-federation.config.ts
@@ -14,6 +14,23 @@ const config: ModuleFederationConfig = {
    * Declare module 'my-external-remote';
    */
   remotes: ['rspack_nx_mf_remote'],
+  shared: (libName) => {
+    const reactShared = [
+      'react',
+      'react-dom',
+      'react/jsx-runtime',
+      'react/jsx-dev-runtime',
+    ];
+    if (reactShared.includes(libName)) {
+      return {
+        singleton: true,
+        version: '18.3.1',
+        requiredVersion: '18.3.1',
+        eager: true,
+      };
+    }
+    return false;
+  },
 };
 
 /**
