+++ libs/vite-plugin-zephyr/src/lib/vite-plugin-zephyr.ts
@@ -31,7 +31,7 @@ function zephyrPlugin(): Plugin {
     resolve_vite_internal_options = resolve;
   });
   let root: string;
-  
+
   let baseHref = '/';
   let mfPlugin: (Plugin & { _options: ModuleFederationOptions }) | undefined;
