+++ libs/zephyr-xpack-internal/src/xpack-extract/mut-webpack-federated-remotes-config.ts
@@ -34,13 +34,12 @@ export function mutWebpackFederatedRemotesConfig<Compiler>(
 
     remoteEntries.forEach((remote) => {
       const [remote_name, remote_version] = remote;
-      const resolved_dep = resolvedDependencyPairs.find(
-        (dep) => {
-          const nameMatch = dep.name === remote_name;
-          const versionMatch = dep.version === 'latest' ? true : dep.version === remote_version;
-          return nameMatch && versionMatch;
-        }
-      );
+      const resolved_dep = resolvedDependencyPairs.find((dep) => {
+        const nameMatch = dep.name === remote_name;
+        const versionMatch =
+          dep.version === 'latest' ? true : dep.version === remote_version;
+        return nameMatch && versionMatch;
+      });
 
       ze_log(`remote_name: ${remote_name}, remote_version: ${remote_version}`);
